rename checkmarx.py -> util.py
